Show correct y-values in the graph, scale according to the game with the highest relative y-value.

+++ langs/java/src/main/java/com/github/tetrisanalyzer/settings/SettingsReader.java
@@ -303,9 +303,9 @@ public class SettingsReader {
             return new Distribution(boardWidth, boardHeight);
         }
         List<String> list = (List)get("distribution");
-        List<Integer> cells = new ArrayList<>();
+        List<Long> cells = new ArrayList<>();
         for (String value : list) {
-            cells.add(Integer.parseInt(value));
+            cells.add(Long.parseLong(value));
         }
         return new Distribution(cells);
     }
